feat: Convert ImageCutout component to vue3
